requirements.txt
##this requirement is for my gcp server which has no cuda
Flask==3.0.0
Flask_Cors==4.0.0
numpy==1.23.5
regex==2017.4.5
Requests==2.31.0
tensorflow-cpu
torch
tqdm==4.64.0
